Fixed batched synchronisers and optimised them a little so that locks are deterministic.

+++ ReactiveTables.Framework/Synchronisation/ReactiveBatchedPassThroughTable.cs
@@ -15,9 +15,12 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 using System.Threading;
 using ReactiveTables.Framework.Columns;
 using ReactiveTables.Framework.Marshalling;
+using ReactiveTables.Framework.Utils;
 
 namespace ReactiveTables.Framework.Synchronisation
 {
@@ -25,12 +28,13 @@ namespace ReactiveTables.Framework.Synchronisation
     {
         private readonly Queue<RowUpdate> _rowUpdatesAdd = new Queue<RowUpdate>();
         private readonly Queue<RowUpdate> _rowUpdatesDelete = new Queue<RowUpdate>();
+        private readonly Dictionary<Type, ITableColumnUpdater> _columnUpdaters = new Dictionary<Type, ITableColumnUpdater>();
         private readonly FieldRowManager _rowManager = new FieldRowManager();
-        private readonly ReactiveTable _targetTable;
+        private readonly IWritableReactiveTable _targetTable;
         private readonly IThreadMarshaller _marshaller;
         private readonly Timer _timer;
         private readonly object _shared = new object();
-        readonly Dictionary<Type, IUpdater> _updaters = new Dictionary<Type, IUpdater>();
+        private readonly System.Timers.Timer _timer1;
 
         public int RowCount
         {
@@ -47,37 +51,81 @@ namespace ReactiveTables.Framework.Synchronisation
             get { throw new NotImplementedException(); }
         }
 
-        public ReactiveBatchedPassThroughTable(ReactiveTable table, IThreadMarshaller marshaller, TimeSpan delay)
+        public Queue<RowUpdate> RowUpdatesAdd
         {
-            _targetTable = table;
+            get
+            {
+                lock (_shared)
+                {
+                    return _rowUpdatesAdd;
+                }
+            }
+        }
+
+        public ReactiveBatchedPassThroughTable(IWritableReactiveTable targetTable, IThreadMarshaller marshaller, TimeSpan delay)
+        {
+            _targetTable = targetTable;
             _marshaller = marshaller;
-            _timer = new Timer(SynchroniseChanges, null, delay, delay);
+//            _timer = new Timer(SynchroniseChanges, null, delay, delay);
+            _timer1 = new System.Timers.Timer(delay.TotalMilliseconds);
+            _timer1.Elapsed += (sender, args) => SynchroniseChanges(null);
+            _timer1.AutoReset = false;
+            _timer1.Start();
         }
 
         private void SynchroniseChanges(object state)
         {
+            // Make copies to control exactly when we lock
+            List<RowUpdate> rowUpdatesAdd = null, rowUpdatesDelete = null;
+            List<ITableColumnUpdater> colUpdaters;
+            lock (_shared)
+            {
+                if (RowUpdatesAdd.Count > 0) rowUpdatesAdd = RowUpdatesAdd.DequeueAllToList();
+                if (_rowUpdatesDelete.Count > 0) rowUpdatesDelete = _rowUpdatesDelete.DequeueAllToList();
+
+                colUpdaters = new List<ITableColumnUpdater>(from u in _columnUpdaters.Values where u.UpdateCount > 0 select u.Clone());
+                foreach (var tableColumnUpdater in _columnUpdaters.Values)
+                {
+                    tableColumnUpdater.Clear();
+                }
+            }
+
+            if (rowUpdatesAdd == null && rowUpdatesDelete == null && colUpdaters.Count == 0)
+            {
+                _timer1.Enabled = true;
+                return;
+            }
+
             // Don't make dispatch granular so that we don't incur as many context switches.
             _marshaller.Dispatch(
                 () =>
                     {
-                        lock (_shared)
+                        try
                         {
-                            while (_rowUpdatesAdd.Count > 0)
+                            if (rowUpdatesAdd != null)
                             {
-                                _rowUpdatesAdd.Dequeue();
-                                _targetTable.AddRow();
+                                for (int i = 0; i < rowUpdatesAdd.Count; i++)
+                                {
+                                    _targetTable.AddRow();
+                                }
                             }
 
-                            foreach (var updater in _updaters)
+                            foreach (var updater in colUpdaters)
                             {
-                                updater.Value.SetValues(_targetTable);
+                                updater.SetValues(_targetTable);
                             }
 
-                            while (_rowUpdatesDelete.Count > 0)
+                            if (rowUpdatesDelete != null)
                             {
-                                var delete = _rowUpdatesDelete.Dequeue();
-                                _targetTable.DeleteRow(delete.RowIndex);
+                                foreach (var delete in rowUpdatesDelete)
+                                {
+                                    _targetTable.DeleteRow(delete.RowIndex);
+                                }
                             }
+                        }
+                        finally
+                        {
+                            _timer1.Enabled = true;
                         }
                     });
         }
@@ -123,16 +171,16 @@ namespace ReactiveTables.Framework.Synchronisation
             BatchedColumnUpdate<T> update = new BatchedColumnUpdate<T> {ColumnId = columnId, Value = value, RowIndex = rowIndex};
             lock (_shared)
             {
-                Updater<T> updater;
+                TableColumnUpdater<T> updater;
                 Type type = typeof (T);
-                if (!_updaters.ContainsKey(type))
+                if (!_columnUpdaters.ContainsKey(type))
                 {
-                    updater = new Updater<T>();
-                    _updaters.Add(type, updater);
+                    updater = new TableColumnUpdater<T>();
+                    _columnUpdaters.Add(type, updater);
                 }
                 else
                 {
-                    updater = (Updater<T>) _updaters[type];
+                    updater = (TableColumnUpdater<T>) _columnUpdaters[type];
                 }
                 updater.Add(update);
             }
@@ -149,7 +197,7 @@ namespace ReactiveTables.Framework.Synchronisation
             RowUpdate update = new RowUpdate(rowIndex, RowUpdate.RowUpdateAction.Add);
             lock (_shared)
             {
-                _rowUpdatesAdd.Enqueue(update);
+                RowUpdatesAdd.Enqueue(update);
             }
             return rowIndex;
         }
@@ -166,7 +214,8 @@ namespace ReactiveTables.Framework.Synchronisation
 
         public void Dispose()
         {
-            _timer.Dispose();
+            if (_timer != null) _timer.Dispose();
+            if (_timer1 != null) _timer1.Dispose();
         }
     }
 
@@ -177,20 +226,34 @@ namespace ReactiveTables.Framework.Synchronisation
         public int RowIndex;
     }
 
-    interface IUpdater
+    interface ITableColumnUpdater
     {
-        void SetValues(ReactiveTable targetTable);
+        void SetValues(IWritableReactiveTable targetTable);
+        ITableColumnUpdater Clone();
+        int UpdateCount { get; }
+        void Clear();
     }
 
-    class Updater<T> : IUpdater
+    class TableColumnUpdater<T> : ITableColumnUpdater
     {
-        readonly Queue<BatchedColumnUpdate<T>> _updates = new Queue<BatchedColumnUpdate<T>>();
+        readonly Queue<BatchedColumnUpdate<T>> _updates;
+
+        public TableColumnUpdater()
+        {
+            _updates = new Queue<BatchedColumnUpdate<T>>();
+        }
+
+        private TableColumnUpdater(IEnumerable<BatchedColumnUpdate<T>> updates)
+        {
+            _updates = new Queue<BatchedColumnUpdate<T>>(updates);
+        }
+
         public void Add(BatchedColumnUpdate<T> update)
         {
             _updates.Enqueue(update);
         }
 
-        public void SetValues(ReactiveTable targetTable)
+        public void SetValues(IWritableReactiveTable targetTable)
         {
             while (_updates.Count > 0)
             {
@@ -198,5 +261,17 @@ namespace ReactiveTables.Framework.Synchronisation
                 targetTable.SetValue(update.ColumnId, update.RowIndex, update.Value);
             }
         }
+
+        public void Clear()
+        {
+            _updates.Clear();
+        }
+
+        public ITableColumnUpdater Clone()
+        {
+            return new TableColumnUpdater<T>(_updates);
+        }
+
+        public int UpdateCount { get { return _updates.Count; } }
     }
 }
